fixed castep and added atom mapping
